refactor(orchestration): extrair modelos de dados para _plan_models.py
Extrai PlanStatus, StepStatus, RiskLevel, PlanStep e ExecutionPlan
(linhas 25-238) de plan_manager.py para o módulo irmão _plan_models.py.

plan_manager.py mantém shim de re-export: todos os imports existentes
continuam resolvendo sem alteração. Singleton get_plan_manager() permanece
em plan_manager.py. _plan_models.py importa all_dependencies_met de _deps
(acíclico) pois ExecutionPlan.get_next_steps o utiliza.

Refs #749
